Split sessions CLI diagnostics
