+++ cmuxTests/CMUXCLISessionsListTests.swift
@@ -1,0 +1,77 @@
+import Foundation
+import Testing
+
+extension CMUXCLIErrorOutputRegressionTests {
+    @Test func testSessionsListReportsCodexIdsMissingFromCodexStore() throws {
+        let cliPath = try bundledCLIPath()
+        let root = FileManager.default.temporaryDirectory
+            .appendingPathComponent("cmux-sessions-list-\(UUID().uuidString)", isDirectory: true)
+        let stateDir = root.appendingPathComponent("state", isDirectory: true)
+        let codexHome = root.appendingPathComponent(".codex", isDirectory: true)
+        try FileManager.default.createDirectory(at: stateDir, withIntermediateDirectories: true)
+        try FileManager.default.createDirectory(at: codexHome, withIntermediateDirectories: true)
+        defer { try? FileManager.default.removeItem(at: root) }
+
+        let sessionId = "019ee74a-3c84-7de3-84f1-ece32f4ecfbb"
+        let workspaceId = "workspace-debug"
+        let surfaceId = "surface-debug"
+        let store: [String: Any] = [
+            "version": 1,
+            "activeSessionsByWorkspace": [
+                workspaceId: [
+                    "sessionId": sessionId,
+                    "updatedAt": 1_781_996_867.0
+                ]
+            ],
+            "activeSessionsBySurface": [
+                surfaceId: [
+                    "sessionId": sessionId,
+                    "updatedAt": 1_781_996_867.0
+                ]
+            ],
+            "sessions": [
+                sessionId: [
+                    "sessionId": sessionId,
+                    "workspaceId": workspaceId,
+                    "surfaceId": surfaceId,
+                    "cwd": "/tmp/cmux/debug",
+                    "startedAt": 1_781_996_800.0,
+                    "updatedAt": 1_781_996_867.0
+                ]
+            ]
+        ]
+        let data = try JSONSerialization.data(withJSONObject: store, options: [.prettyPrinted, .sortedKeys])
+        try data.write(to: stateDir.appendingPathComponent("codex-hook-sessions.json"), options: .atomic)
+
+        var environment = ProcessInfo.processInfo.environment
+        for key in Array(environment.keys) where key.hasPrefix("CMUX_") {
+            environment.removeValue(forKey: key)
+        }
+        environment["CMUX_CLI_SENTRY_DISABLED"] = "1"
+        environment["CMUX_AGENT_HOOK_STATE_DIR"] = stateDir.path
+        environment["CODEX_HOME"] = codexHome.path
+
+        let result = runProcess(
+            executablePath: cliPath,
+            arguments: ["sessions", "list", "--agent", "codex", "--session", sessionId, "--json"],
+            environment: environment,
+            timeout: 5
+        )
+
+        XCTAssertFalse(result.timedOut, result.stdout)
+        XCTAssertEqual(result.status, 0, result.stdout)
+        let outputData = try XCTUnwrap(result.stdout.data(using: .utf8))
+        let object = try XCTUnwrap(JSONSerialization.jsonObject(with: outputData) as? [String: Any])
+        XCTAssertEqual(object["total_matches"] as? Int, 1)
+        let sessions = try XCTUnwrap(object["sessions"] as? [[String: Any]])
+        let session = try XCTUnwrap(sessions.first)
+        XCTAssertEqual(session["session_id"] as? String, sessionId)
+        XCTAssertEqual(session["workspace_id"] as? String, workspaceId)
+        XCTAssertEqual(session["surface_id"] as? String, surfaceId)
+        XCTAssertEqual(session["active_for_workspace"] as? Bool, true)
+        XCTAssertEqual(session["active_for_surface"] as? Bool, true)
+        XCTAssertEqual(session["codex_indexed"] as? Bool, false)
+        XCTAssertEqual(session["codex_transcript_found"] as? Bool, false)
+        XCTAssertEqual(session["session_home"] as? String, codexHome.path)
+    }
+}
